set float16 instead of 32

+++ train.py
@@ -22,7 +22,8 @@ def main(dataset, train_n: int, val_n: int, test_n: int, n_s: int, n_q: int, epo
         WEIGHTSFOLDER / 'best_model.ckpt', verbose=True, mode='min', prefix=dataset
     )
     trainer = pl.Trainer(checkpoint_callback=checkpoint, max_epochs=epochs,
-                         early_stop_callback=early_stop, gpus=gpu, auto_select_gpus=True)
+                         early_stop_callback=early_stop, gpus=gpu,
+                         auto_select_gpus=True, precision=16)
     if not (WEIGHTSFOLDER / 'best_model.ckpt').exists():
         model = PrototypicalNetwork(train_n, test_n, n_s, n_q, lr)
     else:
